basic control of a single neopixel
