+++ apps/lightning/README.md
@@ -1,0 +1,24 @@
+# Lightning
+
+**TODO: Add description**
+
+## Installation
+
+If [available in Hex](https://hex.pm/docs/publish), the package can be installed as:
+
+  1. Add `lightning` to your list of dependencies in `mix.exs`:
+
+    ```elixir
+    def deps do
+      [{:lightning, "~> 0.1.0"}]
+    end
+    ```
+
+  2. Ensure `lightning` is started before your application:
+
+    ```elixir
+    def application do
+      [applications: [:lightning]]
+    end
+    ```
+
