+++ apps/lightning/lib/lightning/colors.ex
@@ -1,0 +1,9 @@
+defmodule Lightning.Colors do
+
+  def red do {255, 0, 0} end
+  def green do {0, 255, 0} end
+  def blue do {0, 0, 255} end
+  def off do {0, 0, 0} end
+  def white do {255, 255, 255} end
+
+end
