+++ apps/lightning/test/lightning_test.exs
@@ -1,0 +1,8 @@
+defmodule LightningTest do
+  use ExUnit.Case
+  doctest Lightning
+
+  test "the truth" do
+    assert 1 + 1 == 2
+  end
+end
