+++ apps/the_eye/lib/the_eye.ex
@@ -7,8 +7,7 @@ defmodule TheEye do
 
   def start(_type, _args) do
     import Supervisor.Spec, warn: false
-
-    neopixel_cfg = Application.get_env(:the_eye, :channel0)
+    neopixel_cfg = Application.get_env(:lightning, :channel0)
 
     # Define workers and child supervisors to be supervised
     children = [
@@ -16,9 +15,6 @@ defmodule TheEye do
       worker(Task, [fn -> init_network() end], restart: :transient, id: Nerves.Init.Network),
       worker(Nerves.Neopixel, [neopixel_cfg, nil]),
     ]
-
-    #For ref
-    # Nerves.Neopixel.render(0, {150, [{0,255,0}]})
 
     # See http://elixir-lang.org/docs/stable/elixir/Supervisor.html
     # for other strategies and supported options
