+++ apps/zeus/README.md
@@ -3,6 +3,7 @@
 To start your Phoenix app:
 
   * Install dependencies with `mix deps.get`
+  * Install Node.js dependencies with `npm install`
   * Start Phoenix endpoint with `mix phoenix.server`
 
 Now you can visit [`localhost:4000`](http://localhost:4000) from your browser.
