+++ apps/zeus/config/prod.exs
@@ -12,7 +12,7 @@ use Mix.Config
 # manifest is generated by the mix phoenix.digest task
 # which you typically run after static files are built.
 config :zeus, Zeus.Endpoint,
-  http: [port: {:system, "PORT"}],
+  http: [port: 80],
   url: [host: "example.com", port: 80],
   cache_static_manifest: "priv/static/manifest.json"
 
@@ -55,10 +55,6 @@ config :logger, level: :info
 #
 #     config :zeus, Zeus.Endpoint, server: true
 #
-# You will also need to set the application root to `.` in order
-# for the new static assets to be served after a hot upgrade:
-#
-#     config :zeus, Zeus.Endpoint, root: "."
 
 # Finally import the config/prod.secret.exs
 # which should be versioned separately.
