+++ apps/zeus/lib/zeus/color_parsing.ex
@@ -2,7 +2,7 @@ defmodule Zeus.ColorParsing do
 
   def parse_rgb_hex(rgb_value) do
     [r1, r2, g1, g2, b1, b2] = rgb_value |> String.codepoints |> Enum.take(-6)
-    [hex_to_number(r1 <> r2), hex_to_number(g1 <> g2), hex_to_number(b1 <> b2)]
+    {hex_to_number(r1 <> r2), hex_to_number(g1 <> g2), hex_to_number(b1 <> b2)}
   end
 
   def hex_to_number(hex_value) do
