+++ apps/zeus/web/static/css/app.css
@@ -1,0 +1,17 @@
+/* This file is for your main application css. */
+
+.color-picker {
+  height: 50px;
+  width: 85px;
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+  margin-bottom: 12px;
+}
+
+input.brightness-picker {
+  width: 100px;
+  margin-left: auto;
+  margin-right: auto;
+  margin-bottom: 12px;
+}
